requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.0
flake8==7.1.1
google-auth==2.37.0
griffe==1.5.1
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
logfire-api==2.7.1
mccabe==0.7.0
openai==1.57.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.10.3
pydantic-ai==0.0.12
pydantic-ai-slim==0.0.12
pydantic_core==2.27.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
